Fix Customer Worker type error
A Customer is reporting an issue where an exception is being thrown when
trying to hatch workers. This is due to the worker count being a
float64, instead of the expected int64.

As we've not been able to replicate the issue with the versions of
ably-boomer and Locust available to us, a decision has been to try and
handle every number type.